requirements.txt
piexif
Pillow
pyexiftool
opencv-python